(#72) 🚀 feat: 이미지 모달 커스텀 훅 제작
